Add files via upload
The given data is for people applying for a loan to a peer-to-peer lending firm. The data contains details of the loan application and eventually how much interest rate was offered to people by the platform. Our solution needs to be able to predict the interest rate which will be offered to people, given their application details as inputs.